# Conflicts:
#	air_ticket_booking/src/main/java/com/example/air_ticket_booking/controller/account/AccountController.java
#	air_ticket_booking/src/main/java/com/example/air_ticket_booking/controller/seat/TypeSeatController.java
#	air_ticket_booking/src/main/java/com/example/air_ticket_booking/repository/route/IRouteRepository.java
#	air_ticket_booking/src/test/java/com/example/air_ticket_booking/controller/account/AccountController_signup.java

+++ air_ticket_booking/src/main/java/com/example/air_ticket_booking/model/ticket/Ticket.java
@@ -12,7 +12,7 @@ public class Ticket {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long idTicket;
-    private Long priceTicket;
+    private double priceTicket;
     private Boolean flagTicket;
     private String namePassenger;
     private Boolean genderPassenger;
@@ -62,11 +62,11 @@ public class Ticket {
         return idTicket;
     }
 
-    public Long getPriceTicket() {
+    public double getPriceTicket() {
         return priceTicket;
     }
 
-    public void setPriceTicket(Long priceTicket) {
+    public void setPriceTicket(double priceTicket) {
         this.priceTicket = priceTicket;
     }
